[xy] Add Postgres source.
