+++ mage_integrations/mage_integrations/sources/postgres/templates/config.json
@@ -1,0 +1,8 @@
+{
+  "database": "",
+  "host": "",
+  "password": "",
+  "port": 5432,
+  "schema": "",
+  "username": ""
+}
