+++ mage_integrations/mage_integrations/sources/sql/utils.py
@@ -1,4 +1,10 @@
-from mage_integrations.sources.constants import COLUMN_TYPE_NULL
+from mage_integrations.sources.constants import (
+    COLUMN_TYPE_BOOLEAN,
+    COLUMN_TYPE_INTEGER,
+    COLUMN_TYPE_NULL,
+    COLUMN_TYPE_NUMBER,
+    COLUMN_TYPE_OBJECT,
+)
 from mage_integrations.utils.array import find
 from typing import Any, Callable, Dict
 
@@ -14,3 +20,16 @@ def build_comparison_statement(
     col_type = column_type_mapping(column_type)
 
     return f"{col} {operator} CAST('{val}' AS {col_type})"
+
+
+def column_type_mapping(column_type: str) -> str:
+    if COLUMN_TYPE_BOOLEAN == column_type:
+        return 'BOOL'
+    elif COLUMN_TYPE_INTEGER == column_type:
+        return 'UNSIGNED INTEGER'
+    elif COLUMN_TYPE_NUMBER == column_type:
+        return 'DECIMAL'
+    elif COLUMN_TYPE_OBJECT == column_type:
+        return 'TEXT'
+
+    return 'CHAR'
